changed files to tsx, added types to remove error
