+++ react-carousel-component/my-app/src/Banner.tsx
@@ -1,0 +1,11 @@
+type Props = {
+  image: string;
+};
+
+export default function Banner({ image }: Props) {
+  return (
+    <div className="row align-center justify-right image">
+      <img src={`/images/${image}.png`} alt={`pokemon id ${image}`} />
+    </div>
+  );
+}
