+++ react-carousel-component/my-app/src/NextButton.tsx
@@ -1,10 +1,10 @@
 import { FaAngleRight } from 'react-icons/fa';
 
-type props = {
-  onClick: Function,
+type Props = {
+  onClick: () => void;
 };
 
-export default function NextButton({ onClick }: props) {
+export default function NextButton({ onClick }: Props) {
   return (
     <div className="row justify-center arrow-right">
       <FaAngleRight className="fa-right" onClick={onClick} />
